Unify Android notification API
